// antigo/app-admin-next/models/SemcompApiModels.ts
export type SemcompApiAdminUser = {
  id: string;
  email: string;
  adminRole: string[];
  createdAt: number;
  updatedAt: number;
};